travis: setup automatic updates and releases
Update the Travis file to add cron triggered updates and releases of
citeproc-py-styles. For this, we added run-update.sh and an update stage
to Travis.

run-update.sh triggers a git submodule update for the CSL Style
repository. If there are changes in this repository we commit these
changes and create a tagged commit with a new release. To have simple
and reliable versioning we switch away from semantic versioning to date
versioning.

The Travis file is changed to distinguish three different situations:
1. Regular build
2. Cron triggered build
3. Tagged build

For the regular builds, we don't change a lot. We added python 3.6
support, this in line with other Invenio modules. As a result, this
build triggers the test for three different Python versions: 2.7, 3.5
and 3.6.

The cron triggered builds are the new update builds. The cron triggered
builds only run the update stage. This stage updates the submodule to
the latest version and, if there are changes, tags a new release. After
a simple test (just Python 3.6) these changes are pushed.

The tagged builds run two stages. The first stage is a regular build,
testing the module for different Python versions. If this succeeds the
release stage is triggered, this stage releases the new package to Pypi.